Stop deleting observability user during upgrades
Adjust the post-uninstallation phase so that the observability user is
only deleted during full uninstallation. This prevents the user from
being deleted during upgrades and ensures the service contines to work
as intended after a major rhel upgrade. Also add precautionary runtime
requirements for shadow-utils and util-linux as both are required for
the creation of the observability user.

Signed-off-by: ccowman <ccowman@redhat.com>